Never return the student from a swipe
Only the list of students, which is much much faster. The student does
not need to be cached on a swipe, this is a masssive waste of effort.
When the user needs to open a student, fetch the data then. Otherwise,
only fetch exactly what is needed at the time of use.